requirements.txt
# Core ML/AI packages
torch
torchaudio
numpy

# Speech Recognition (Whisper)
faster-whisper

# Text-to-Speech (Piper)
piper-tts

# Voice Activity Detection
silero-vad

# ONNX Runtime (GPU support for CUDA)
onnxruntime-gpu

# Audio Processing
sounddevice
scipy
pyrubberband

# GUI (PyQt6)
PyQt6

# API Server
Flask

# Utilities
keyboard